+++ frontend/src/containers/HomePage.js
@@ -3,6 +3,6 @@ import { connect } from 'react-redux'
 import { HomePage } from 'components'
 
 
-const mapStateToProps = (state) => ({user_state: state.ps4u.user_state})
+const mapStateToProps = (state) => ({user_state: state.users.user_state})
 const mapDispatchToProps = (dispatch) => ({})
 export default connect(mapStateToProps, mapDispatchToProps)((props)=>(<HomePage {...props} />))
